# 1203/demos/userapp/openapi.yaml
openapi: 3.0.3
info:
  title: UserApp API
  description: REST API für Benutzerverwaltung mit Authentifizierung, Profilverwaltung und Admin-Funktionen.
  version: 1.0.0
  contact:
    name: UserApp Team
servers:
  - url: https://api.userapp.example.com/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─────────────────────────────────────────────
  # Reusable Schemas
  # ─────────────────────────────────────────────
  schemas:
...
    Role:
      type: string
      enum: [USER, ADMIN]
      example: USER
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Beschreibung des Fehlers"
...
  responses:
    Unauthorized:
      description: Ungültige oder fehlende Anmeldedaten
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 401
            error: Unauthorized
            message: "Ungültige E-Mail-Adresse oder Passwort."
    Forbidden:
      description: Zugriff verweigert
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 403
            error: Forbidden
            message: "Keine Berechtigung für diese Aktion."
...
    UnprocessableEntity:
      description: Validierungsfehler
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 422
            error: Unprocessable Entity
            message: "Das Passwort erfüllt nicht die Mindestanforderungen."

# ─────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Neues Konto registrieren
      description: Erstellt ein neues Benutzerkonto. Das Passwort wird mit bcrypt (≥12 Runden) gehasht. Nach der Registrierung wird eine Verifizierungs-E-Mail versendet (TTL 24 h).
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, password]
              properties:
                firstName:
                  type: string
                  maxLength: 50
                  example: "Max"
                lastName:
                  type: string
                  maxLength: 50
                  example: "Mustermann"
                email:
                  type: string
                  format: email
                  example: "max.mustermann@example.com"
                password:
                  type: string
                  minLength: 8
                  description: "Mind. 8 Zeichen, inkl. Großbuchstabe, Zahl und Sonderzeichen"
                  example: "Sicher!123"
      responses:
        "201":
          description: Konto erfolgreich erstellt
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                    format: email
                  role:
                    $ref: "#/components/schemas/Role"
        "400":
          description: Pflichtfelder fehlen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: E-Mail-Adresse bereits vergeben
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                statusCode: 409
                error: Conflict
                message: "Diese E-Mail-Adresse ist bereits registriert."
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ── 2. Login ──────────────────────────────
  /auth/login:
    post:
      tags: [Auth]
      summary: Anmelden
      description: Authentifiziert einen Nutzer und gibt Access- und Refresh-Token zurück. Bei 5 Fehlversuchen pro IP wird der Zugang für 10 Minuten gesperrt.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "max.mustermann@example.com"
                password:
                  type: string
                  example: "Sicher!123"
      responses:
        "200":
          description: Erfolgreich angemeldet
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: "JWT (TTL: 15 Minuten)"
                    example: "eyJhbGciOiJSUzI1NiJ9..."
                  refreshToken:
                    type: string
                    description: "Opaque Token (TTL: 7 Tage)"
                    example: "dGhpcyBpcyBhIHJlZnJlc2g..."
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── 5b. Passwort vergessen ─────────────────
...